Updated command to iniitialize CMIP6 model seeding config.  Added command to write CMIP^ bash vars.  Updated vocabs shell file.

+++ bash/cmip6/models/push_seeding_config.sh
@@ -6,14 +6,18 @@ source $ESDOC_HOME/bash/utils.sh
 # Main entry point.
 main()
 {
-	log "GITHUB : writing CMIP6 model seeding configuration ..."
+	log "GITHUB : writing CMIP6 model seeding configuration:"
 
 	declare DEST=$ESDOC_HOME/repos/institutional
 
 	for institution_id in "${INSTITUTION_ID[@]}"
 	do
 		mkdir -p $DEST/$institution_id/cmip6/models
-		cp $ESDOC_HOME/repos/esdoc-docs/cmip6/models/seeding/$institution_id.json $DEST/$institution_id/cmip6/models/model-defaults.json
+		declare target=$ESDOC_HOME/repos/esdoc-docs/cmip6/models/seeding/$institution_id.json
+		if [ -f "$target" ]; then
+			cp $ESDOC_HOME/repos/esdoc-docs/cmip6/models/seeding/$institution_id.json $DEST/$institution_id/cmip6/models/model-defaults.json
+			log "... "$institution_id
+		fi
 	done
 
 	log "GITHUB : CMIP6 model seeding configuration written..."
